pyproject.toml
[tool.poetry]
name = "study-case"
version = "0.1.0"
description = "Data Engineering case study"
authors = ["Pedro Aguila Godoy <pedroaguilag@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
matplotlib = "^3.9.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
